refactor: move shared methods into their own module
This is for call pattern and call history report

Refs: BLAIS5-3036
